requirements.txt
lmdb==1.6.2
matplotlib==3.10.3
mc==0.0
numpy==2.3.0
opencv_python==4.10.0.84
phantominator==0.7.0
Pillow==11.2.1
scipy==1.15.3
skimage==0.0
tabulate==0.9.0
torch==2.3.1+cu121
torchvision==0.18.1+cu121
